+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### HEAD
 
+* Update Bower dependencies
+* Update npm module dependencies
+* Add AUTHORS file instead of specifiying it in package.json
+* Add grunt-imageoptim as task
+
 * Update jQuery to 1.10.2
 * Use SASS-Mixins via Bower
 * Use Bower for package management and replace vendor packages (#24)